+++ src/components/DFR3Viewer.jsx
@@ -454,6 +454,10 @@ class DFR3Viewer extends React.Component {
 			else if (curve.className === "ParametricFragilityCurve"){
 				plotData = chartSampler.computeParametricSampes(0.001, 1, 1000, curve.curveType, curve.parameters);
 			}
+			else if (curve.className === "PeriodBuildingFragilityCurve"){
+				plotData = chartSampler.computePeriodBuildingSamples(0, 5, 1000, curve.fsParam0, curve.fsParam1,
+					curve.fsParam2, curve.fsParam3, curve.fsParam4, curve.fsParam5)
+			}
 			// TODO if curve type none of the above
 			// TODO need to have a widget to display the error in a uniformed way